oxjsonld

Enum JsonLdErrorCode

#[non_exhaustive]
pub enum JsonLdErrorCode {

    CollidingKeywords,
    ConflictingIndexes,
    ContextOverflow,
    CyclicIriMapping,
    InvalidIdValue,
    InvalidImportValue,
    InvalidIncludedValue,
    InvalidIndexValue,
    InvalidNestValue,
    InvalidPrefixValue,
    InvalidPropagateValue,
    InvalidProtectedValue,
    InvalidReverseValue,
    InvalidVersionValue,
    InvalidBaseDirection,
    InvalidBaseIri,
    InvalidContainerMapping,
    InvalidContextEntry,
    InvalidContextNullification,
    InvalidDefaultLanguage,
    InvalidIriMapping,
    InvalidJsonLiteral,
    InvalidKeywordAlias,
    InvalidLanguageMapValue,
    InvalidLanguageMapping,
    InvalidLanguageTaggedString,
    InvalidLanguageTaggedValue,
    InvalidLocalContext,
    InvalidRemoteContext,
    InvalidReverseProperty,
    InvalidReversePropertyMap,
    InvalidReversePropertyValue,
    InvalidScopedContext,
    InvalidSetOrListObject,
    InvalidStreamingKeyOrder,
    InvalidTermDefinition,
    InvalidTypeMapping,
    InvalidTypeValue,
    InvalidTypedValue,
    InvalidValueObject,
    InvalidValueObjectValue,
    InvalidVocabMapping,
    IriConfusedWithPrefix,
    KeywordRedefinition,
    LoadingDocumentFailed,
    LoadingRemoteContextFailed,
    ProcessingModeConflict,
    ProtectedTermRedefinition,

}

A JSON-LD error code

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

CollidingKeywords

Two properties which expand to the same keyword have been detected. This might occur if a keyword and an alias thereof are used at the same time.

ConflictingIndexes

Multiple conflicting indexes have been found for the same node.

ContextOverflow

Maximum number of @context URLs exceeded.

CyclicIriMapping

A cycle in IRI mappings has been detected.

InvalidIdValue

An @id entry was encountered whose value was not a string.

InvalidImportValue

An invalid value for @import has been found.

InvalidIncludedValue

An included block contains an invalid value.

InvalidIndexValue

An @index entry was encountered whose value was not a string.

InvalidNestValue

An invalid value for @nest has been found.

InvalidPrefixValue

An invalid value for @prefix has been found.

InvalidPropagateValue

An invalid value for @propagate has been found.

InvalidProtectedValue

An invalid value for @protected has been found.

InvalidReverseValue

An invalid value for an @reverse entry has been detected, i.e., the value was not a map.

InvalidVersionValue

The @version entry was used in a context with an out of range value.

InvalidBaseDirection

The value of @direction is not โltrโ, โrtlโ, or null and thus invalid.

InvalidBaseIri

An invalid base IRI has been detected, i.e., it is neither an IRI nor null.

InvalidContainerMapping

An @container entry was encountered whose value was not one of the following strings: @list, @set, @language, @index, @id, @graph, or @type.

InvalidContextEntry

An entry in a context is invalid due to processing mode incompatibility.

InvalidContextNullification

An attempt was made to nullify a context containing protected term definitions.

InvalidDefaultLanguage

The value of the default language is not a string or null and thus invalid.

InvalidIriMapping

A local context contains a term that has an invalid or missing IRI mapping.

InvalidJsonLiteral

An invalid JSON literal was detected.

InvalidKeywordAlias

An invalid keyword alias definition has been encountered.

InvalidLanguageMapValue

An invalid value in a language map has been detected. It MUST be a string or an array of strings.

InvalidLanguageMapping

An @language entry in a term definition was encountered whose value was neither a string nor null and thus invalid.

InvalidLanguageTaggedString

A language-tagged string with an invalid language value was detected.

InvalidLanguageTaggedValue

A number, true, or false with an associated language tag was detected.

InvalidLocalContext

An invalid local context was detected.

InvalidRemoteContext

No valid context document has been found for a referenced remote context.

InvalidReverseProperty

An invalid reverse property definition has been detected.

InvalidReversePropertyMap

An invalid reverse property map has been detected. No keywords apart from @context are allowed in reverse property maps.

InvalidReversePropertyValue

An invalid value for a reverse property has been detected. The value of an inverse property must be a node object.

InvalidScopedContext

The local context defined within a term definition is invalid.

InvalidSetOrListObject

A set object or list object with disallowed entries has been detected.

InvalidStreamingKeyOrder

The key ordering is not compatible with the streaming profile.

InvalidTermDefinition

An invalid term definition has been detected.

InvalidTypeMapping

An @type entry in a term definition was encountered whose value could not be expanded to an IRI.

InvalidTypeValue

An invalid value for an @type entry has been detected, i.e., the value was neither a string nor an array of strings.

InvalidTypedValue

A typed value with an invalid type was detected.

InvalidValueObject

A value object with disallowed entries has been detected.

InvalidValueObjectValue

An invalid value for the @value entry of a value object has been detected, i.e., it is neither a scalar nor null.

InvalidVocabMapping

An invalid vocabulary mapping has been detected, i.e., it is neither an IRI nor null.

IriConfusedWithPrefix

When compacting an IRI would result in an IRI which could be confused with a compact IRI (because its IRI scheme matches a term definition and it has no IRI authority).

KeywordRedefinition

A keyword redefinition has been detected.

LoadingDocumentFailed

The document could not be loaded or parsed as JSON.

LoadingRemoteContextFailed

There was a problem encountered loading a remote context.

ProcessingModeConflict

An attempt was made to change the processing mode which is incompatible with the previous specified version.

ProtectedTermRedefinition

An attempt was made to redefine a protected term.